requirements.txt
streamlit
tweepy
python-dotenv
streamlit-authenticator
Pyrebase4
setuptools